+++ BraintreeCore/BTAPIClient_Internal.h
@@ -27,15 +27,19 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, strong) BTAnalyticsService *analyticsService;
 
 /**
- @brief Analytics should only be posted by internal clients.
+ @brief Sends this event and all queued analytics events. Use `queueAnalyticsEvent` for low priority events.
 */
 - (void)sendAnalyticsEvent:(NSString *)eventName;
+
+/**
+ @brief Queues an analytics event to be sent.
+ */
+- (void)queueAnalyticsEvent:(NSString *)eventName;
 
 /**
  @brief An internal initializer to toggle whether to send an analytics event during initialization.
  @discussion This prevents copyWithSource:integration: from sending a duplicate event. It can also be used to suppress excessive network chatter during testing.
 */
-
 - (nullable instancetype)initWithAuthorization:(NSString *)authorization sendAnalyticsEvent:(BOOL)sendAnalyticsEvent;
 
 @end