+++ UnitTests/BTAPIClient_SwiftTests.swift
@@ -75,7 +75,7 @@ class BTAPIClient_SwiftTests: XCTestCase {
         apiClient.configurationHTTP = mockHTTP
        
         let expectation = self.expectation(description: "Callback invoked")
-        apiClient.fetchOrReturnRemoteConfiguration() { _ in
+        apiClient.fetchOrReturnRemoteConfiguration() { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/configuration")
             XCTAssertEqual(mockHTTP.lastRequestParameters?["configVersion"] as? String, "3")
             expectation.fulfill()
@@ -93,9 +93,9 @@ class BTAPIClient_SwiftTests: XCTestCase {
         apiClient.http = mockHTTP
        
         var expectation = self.expectation(description: "Callback invoked")
-        apiClient.fetchPaymentMethodNonces() { _ in
+        apiClient.fetchPaymentMethodNonces() { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/payment_methods")
-            XCTAssertFalse(mockHTTP.lastRequestParameters!["default_first"] as! Bool)
+            XCTAssertEqual(mockHTTP.lastRequestParameters!["default_first"] as? String, "false")
             XCTAssertEqual(mockHTTP.lastRequestParameters!["session_id"] as? String, apiClient.metadata.sessionId)
             expectation.fulfill()
         }
@@ -103,18 +103,18 @@ class BTAPIClient_SwiftTests: XCTestCase {
         waitForExpectations(timeout: 1, handler: nil)
        
         expectation = self.expectation(description: "Callback invoked")
-        apiClient.fetchPaymentMethodNonces(true) { _ in
+        apiClient.fetchPaymentMethodNonces(true) { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/payment_methods")
-            XCTAssertTrue(mockHTTP.lastRequestParameters!["default_first"] as! Bool)
+            XCTAssertEqual(mockHTTP.lastRequestParameters!["default_first"] as? String, "true")
             expectation.fulfill()
         }
         
         waitForExpectations(timeout: 1, handler: nil)
         
         expectation = self.expectation(description: "Callback invoked")
-        apiClient.fetchPaymentMethodNonces(false) { _ in
+        apiClient.fetchPaymentMethodNonces(false) { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/payment_methods")
-            XCTAssertFalse(mockHTTP.lastRequestParameters!["default_first"] as! Bool)
+            XCTAssertEqual(mockHTTP.lastRequestParameters!["default_first"] as? String, "false")
             expectation.fulfill()
         }
         
@@ -205,13 +205,16 @@ class BTAPIClient_SwiftTests: XCTestCase {
         let mockHTTP = BTFakeHTTP()!
         mockHTTP.stubRequest("GET", toEndpoint: "/client_api/v1/payment_methods", respondWith: [], statusCode: 200)
         apiClient.http = mockHTTP
+        let mockConfigurationHTTP = BTFakeHTTP()!
+        mockConfigurationHTTP.stubRequest("GET", toEndpoint: "/client_api/v1/configuration", respondWith: [], statusCode: 200)
+        apiClient.configurationHTTP = mockConfigurationHTTP
 
         XCTAssertEqual((apiClient.clientToken!.json["version"] as! BTJSON).asIntegerOrZero(), 3)
 
         var expectation = self.expectation(description: "Callback invoked")
-        apiClient.fetchPaymentMethodNonces() { _ in
+        apiClient.fetchPaymentMethodNonces() { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/payment_methods")
-            XCTAssertFalse(mockHTTP.lastRequestParameters!["default_first"] as! Bool)
+            XCTAssertEqual(mockHTTP.lastRequestParameters!["default_first"] as? String, "false")
             XCTAssertEqual(mockHTTP.lastRequestParameters!["session_id"] as? String, apiClient.metadata.sessionId)
             expectation.fulfill()
         }
@@ -219,18 +222,18 @@ class BTAPIClient_SwiftTests: XCTestCase {
         waitForExpectations(timeout: 1, handler: nil)
 
         expectation = self.expectation(description: "Callback invoked")
-        apiClient.fetchPaymentMethodNonces(true) { _ in
+        apiClient.fetchPaymentMethodNonces(true) { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/payment_methods")
-            XCTAssertTrue(mockHTTP.lastRequestParameters!["default_first"] as! Bool)
+            XCTAssertEqual(mockHTTP.lastRequestParameters!["default_first"] as? String, "true")
             expectation.fulfill()
         }
 
         waitForExpectations(timeout: 1, handler: nil)
 
         expectation = self.expectation(description: "Callback invoked")
-        apiClient.fetchPaymentMethodNonces(false) { _ in
+        apiClient.fetchPaymentMethodNonces(false) { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/payment_methods")
-            XCTAssertFalse(mockHTTP.lastRequestParameters!["default_first"] as! Bool)
+            XCTAssertEqual(mockHTTP.lastRequestParameters!["default_first"] as? String, "false")
             expectation.fulfill()
         }