+++ UnitTests/UnitTests-Bridging-Header.h
@@ -2,6 +2,7 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
+#import "BraintreeAmericanExpress.h"
 #import "BraintreeCore.h"
 #import "BraintreeCard.h"
 #import "BraintreeApplePay.h"
@@ -34,6 +35,7 @@
 #import "Braintree-Version.h"
 #import "PPDataCollector_Internal.h"
 #import "BTDropInUtil.h"
+#import "BTAmericanExpressClient_Internal.h"
 
 #import "BTSpecHelper.h"
 #import <OCMock/OCMock.h>